src/requirements.txt
django==2.0.6
gunicorn==19.8.1
djangorestframework==3.8.2
django-environ==0.4.5
psycopg2-binary==2.7.5
djangorestframework-jwt==1.11.0
